refactor: model Context에서 제외, 기존 Context Command화
