+++ src/App.tsx
@@ -1,15 +1,15 @@
 import Canvas from '@/views/Canvas';
 import PropertiesPanel from '@/views/PropertiesPanel';
 import ToolBar from '@/views/ToolBar';
-import ContextProvider from '@/models/ContextProvider';
+import CommandProvider from '@/commands/CommandProvider';
 
 const App = () => {
   return (
-    <ContextProvider>
+    <CommandProvider>
       <ToolBar />
       <Canvas />
       <PropertiesPanel />
-    </ContextProvider>
+    </CommandProvider>
   );
 };
