+++ src/models/GraphicObjectInterface.ts
