+++ src/views/PropertiesPanel/Layers/index.tsx
@@ -2,21 +2,21 @@ import { useContext, useState } from 'react';
 
 import s from './Layers.module.css';
 
-import useSubscribe from '@/libs/hooks/useSubscribe';
+import useSubscribe from '@/hooks/useSubscribe';
 import {
-  ControllerContext,
+  CommandContext,
   SelectedObjectsContext,
-} from '@/models/GraphicEditorContext';
+} from '@/commands/GraphicEditorContext';
 
 const Layers = () => {
   const objects = useSubscribe();
   const selectedObjects = useContext(SelectedObjectsContext);
-  const controller = useContext(ControllerContext);
+  const command = useContext(CommandContext);
 
   const [draggedItem, setDraggedItem] = useState<string | null>(null);
   const [dragOverIndex, setDragOverIndex] = useState<number | null>(null);
 
-  const { reorderLayers, select } = controller;
+  const { reorderLayers, select } = command;
 
   const handleDragStart = (id: string) => {
     setDraggedItem(id);
