+++ src/views/ToolBar/index.tsx
@@ -1,20 +1,20 @@
-import { ControllerContext } from '@/models/GraphicEditorContext';
+import { CommandContext } from '@/commands/GraphicEditorContext';
 import { useContext } from 'react';
 
 import s from './ToolBar.module.css';
 
 const ToolBar = () => {
-  const controller = useContext(ControllerContext);
+  const command = useContext(CommandContext);
 
   return (
     <div className={s.Wrapper}>
-      <button onClick={() => controller.add('rectangle')}>Rectangle</button>
-      <button onClick={() => controller.add('ellipse')}>Ellipse</button>
-      <button onClick={() => controller.add('line')}>Line</button>
-      <button onClick={() => controller.add('image')}>Image</button>
-      <button onClick={() => controller.add('text')}>Text</button>
-      <button onClick={() => controller.remove()}>Delete</button>
-      <button onClick={() => controller.clear()}>Clear Canvas</button>
+      <button onClick={() => command.add('rectangle')}>Rectangle</button>
+      <button onClick={() => command.add('ellipse')}>Ellipse</button>
+      <button onClick={() => command.add('line')}>Line</button>
+      <button onClick={() => command.add('image')}>Image</button>
+      <button onClick={() => command.add('text')}>Text</button>
+      <button onClick={() => command.remove()}>Delete</button>
+      <button onClick={() => command.clear()}>Clear Canvas</button>
     </div>
   );
 };
